# Conflicts:
#	app/src/main/java/fr/univorleans/etu/renardez/kodapp/MainActivity.java

+++ app/src/main/java/fr/univorleans/etu/renardez/kodapp/entities/daos/PositionUserDao.java
@@ -17,7 +17,7 @@ public interface PositionUserDao {
     List<PositionUser> getAllPU();
 
     @Query("SELECT * FROM positions WHERE id = :ids")
-    List<PositionUser> getSomePu(int... ids);
+    List<PositionUser> getSomePu(long... ids);
 
     @Insert
     long insertPos(PositionUser positionUser);
